split xo-lib in multiple modules
