+++ packages/vhd-lib/src/_constant-stream.js
