+++ packages/xo-vmdk-to-vhd/src/index.js
@@ -1,11 +1,11 @@
-import { createReadableRawVHDStream } from 'vhd-lib'
+import { createReadableRawStream } from 'vhd-lib'
 
 import { VMDKDirectParser } from './vmdk-read'
 
 async function convertFromVMDK (vmdkReadStream) {
   const parser = new VMDKDirectParser(vmdkReadStream)
   const header = await parser.readHeader()
-  return createReadableRawVHDStream(header.capacitySectors * 512, parser)
+  return createReadableRawStream(header.capacitySectors * 512, parser)
 }
 
 export { convertFromVMDK as default }
